Moved to static libraries for submodules

+++ scripts/build-pip.sh
@@ -60,7 +60,11 @@ fi
 # bdist_wheel should auto-detect platform from binary files, but we can also specify it
 rm -rf dist/* wheelhouse/*
 python setup.py bdist_wheel
-auditwheel repair dist/moonshine_voice-*.whl -w dist/
-rm -rf dist/moonshine_voice-*-linux_*.whl
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+	auditwheel repair dist/moonshine_voice-*.whl -w dist/
+	rm -rf dist/moonshine_voice-*-linux_*.whl
+fi
 
-twine upload dist/*
+if [[ "$1" == "upload" ]]; then
+	twine upload dist/*
+fi
